@@ -24,67 +24,23 @@
 
 package org.prowl.aprslib.parser;
 
-public enum APRSTypes implements java.io.Serializable {
-   /**
-    * Item type not specified
-    */
-   UNSPECIFIED,
-   /**
-    * Position Report
-    */
-   POSITION,
-   /**
-    * Weather Station
-    */
-   WX,
-   /**
-   *
-   */
-   THIRDPARTY,
-   /**
-   *
-   */
-   QUERY,
-   /**
-   *
-   */
-   OBJECT,
-   /**
-   *
-   */
-   ITEM,
-   /**
-   *
-   */
-   NORMAL,
-   /**
-   *
-   */
-   KILL,
-   /**
-   *
-   */
-   STATUS,
-   /**
-   *
-   */
-   STATCAPA,
-   /**
-    * Telemetry report format
-    */
-   TELEMETRY,
-   /**
-   *
-   */
-   USERDEF,
-   /**
-   *
-   */
-   MESSAGE,
-   /**
-   *
-   */
-   NWS; // Used on fap.getSubtype()
+public enum APRSTypes {
+    T_UNSPECIFIED,
+	T_TIMESTAMP,
+	T_POSITION,
+	T_WX,
+	T_THIRDPARTY,
+	T_QUERY,
+	T_OBJECT,
+	T_ITEM,
+	T_NORMAL,
+	T_KILL,
+	T_STATUS,
+	T_STATCAPA,
+	T_TELEMETRY,
+	T_USERDEF,
+	T_MESSAGE,
+	T_NWS;  // Used on fap.getSubtype()
 
-   ;
+	;
 }
