@@ -20,77 +20,79 @@
  */
 package org.prowl.aprslib.parser;
 
-import java.io.Serializable;
 import java.util.ArrayList;
 
 /**
- * This class represents a single digipeater in a TNC2-format VIA string.
- *
+ * 
  * @author johng
+ * This class represents a single digipeater in a TNC2-format VIA string.
+ * 
  */
-public class Digipeater extends Callsign implements Serializable {
-   private static final long serialVersionUID = 1L;
-   private boolean           used;
+public class Digipeater extends Callsign {
+	private static final long serialVersionUID = 1L;
+    private boolean used;
+    
+    public Digipeater(String call) {
+	super(call.replaceAll("\\*", ""));
+        if ( call.indexOf("*") >= 0 ) {
+		setUsed(true);
+	}
+    }
+    public Digipeater(byte[] data, int offset) {
+	    super(data, offset);
+	    this.used = (data[offset + 6] & 0x80) == 0x80;
+    }
 
-   public Digipeater(String call) {
-      super(call.replaceAll("\\*", ""));
-      if (call.indexOf("*") >= 0) {
-         setUsed(true);
-      }
-   }
+    /** parse a comma-separated list of digipeaters
+     * @return the list of digipeaters as an array
+     */
+    public static ArrayList<Digipeater> parseList(String digiList, boolean includeFirst) {
+	String[] digiTemp = digiList.split(",");
+	ArrayList<Digipeater> digis = new ArrayList<Digipeater>();
+	boolean includeNext = includeFirst;
+	// for now, '*' is set for all digis with used bit.
+	// however, only the last used digi should have a '*'
+	for (String digi : digiTemp) {
+		String digiTrim = digi.trim();
+		if (digiTrim.length() > 0 && includeNext)
+			digis.add(new Digipeater(digiTrim));
+		includeNext = true;
+	}
+	return digis;
+    }
 
-   public Digipeater(byte[] data, int offset) {
-      super(data, offset);
-      this.used = (data[offset + 6] & 0x80) == 0x80;
-   }
+    /**
+     * @return the used
+     */
+    public boolean isUsed() {
+        return used;
+    }
 
-   /**
-    * parse a comma-separated list of digipeaters
-    *
-    * @param digiList
-    *           the list of digis
-    * @param includeFirst
-    *           includes the first call
-    * @return the list of digipeaters as an array
-    */
-   public static ArrayList<Digipeater> parseList(String digiList, boolean includeFirst) {
-      String[] digiTemp = digiList.split(",");
-      ArrayList<Digipeater> digis = new ArrayList<Digipeater>();
-      int first = includeFirst ? 0 : 1;
-      // for now, '*' is set for all digis with used bit.
-      // however, only the last used digi should have a '*'
-      if (digiTemp.length >= first) {
-         for (int i = first; i < digiTemp.length; i++) {
-            digis.add(new Digipeater(digiTemp[i]));
-         }
-      }
-      return digis;
-   }
+    /**
+     * @param used the used to set
+     */
+    public void setUsed(boolean used) {
+        this.used = used;
+    }
+    
+    
+    /** 
+     * @return String
+     */
+    @Override
+    public String toString() {
+        return super.toString() + ( isUsed() ? "*":"");
+    }
 
-   /**
-    * @return the used
-    */
-   public boolean isUsed() {
-      return used;
-   }
-
-   /**
-    * @param used
-    *           the used to set
-    */
-   public void setUsed(boolean used) {
-      this.used = used;
-   }
-
-   @Override
-   public String toString() {
-      return super.toString() + (isUsed() ? "*" : "");
-   }
-
-   @Override
-   public byte[] toAX25() throws IllegalArgumentException {
-      byte[] ax25 = super.toAX25();
-      ax25[6] |= (isUsed() ? 0x80 : 0);
-      return ax25;
-   }
+    
+    /** 
+     * @return byte[]
+     * @throws IllegalArgumentException
+     */
+    @Override
+    public byte[] toAX25() throws IllegalArgumentException {
+        byte[] ax25 = super.toAX25();
+	ax25[6] |= (isUsed()?0x80:0);
+	return ax25;
+    }
 }
