@@ -20,153 +20,146 @@
  */
 package org.prowl.aprslib.parser;
 
-import java.io.Serializable;
+public class MessagePacket extends APRSData {
+	private static final long serialVersionUID = 1L;
+    private String messageBody;
+    private String messageNumber;
+    private String targetCallsign ="";
+    private boolean isAck = false;
+    private boolean isRej = false;
+    
+    public MessagePacket( byte[] bodyBytes, String destCall ) {
+        super(bodyBytes);
+        String message = new String(bodyBytes);
+        if ( message.length() < 2) {
+            this.hasFault = true;
+            return;
+        }
+        int msgSpc = message.indexOf(':', 2);
+        if ( msgSpc < 1 ) {
+        	this.targetCallsign = "UNKNOWN";
+        } else {
+        	targetCallsign = message.substring(1,msgSpc).trim().toUpperCase();
+        }
+        int msgNumberIdx = message.lastIndexOf('{');
+        this.messageNumber="";
+        if ( msgNumberIdx > -1 ) {
+            this.messageNumber = message.substring(msgNumberIdx+1);
+            messageBody = message.substring(11,msgNumberIdx);
+        } else {
+            messageBody = message.substring(11);
+        }
+        String lcMsg = messageBody.toLowerCase();
+        if ( lcMsg.startsWith("ack") ) {
+        	isAck = true;
+        	this.messageNumber = messageBody.substring(3,messageBody.length());
+		this.messageBody = messageBody.substring(0, 3);
+        }
+        if ( lcMsg.startsWith("rej") ) {
+        	isRej = true;
+        	this.messageNumber = messageBody.substring(3,messageBody.length());
+		this.messageBody = messageBody.substring(0, 3);
+        }
+    }
+    
+    public MessagePacket(String targetCallsign, String messageBody, String messageNumber) {
+    	this.messageBody = messageBody;
+    	this.targetCallsign = targetCallsign;
+    	this.messageNumber = messageNumber;
+    	if ( messageBody.equals("ack") ) isAck = true;
+    	if ( messageBody.equals("rej") ) isRej = true;
+    	//super.setDataTypeIdentifier(':');
+    }
+    
+    /**
+     * @return the messageBody
+     */
+    public String getMessageBody() {
+        return this.messageBody;
+    }
 
-public class MessagePacket extends InformationField implements Serializable {
-   private static final long serialVersionUID = 1L;
-   private String            messageBody;
-   private String            messageNumber;
-   private String            targetCallsign   = "";
-   private boolean           isAck            = false;
-   private boolean           isRej            = false;
+    /**
+     * @param messageBody the messageBody to set
+     */
+    public void setMessageBody(String messageBody) {
+        this.messageBody = messageBody;
+    }
 
-   public MessagePacket(byte[] bodyBytes, String destCall) {
-      super(bodyBytes);
-      String message = new String(bodyBytes);
-      if (destCall.toUpperCase().equals("BEACON")) {
-         this.messageNumber = "";
-         this.messageBody = message.substring(11);
-         this.targetCallsign = "BEACON";
-         return;
-      }
-      if (message.length() < 2) {
-         this.hasFault = true;
-         return;
-      }
-      int msgSpc = message.indexOf(':', 2);
-      if (msgSpc < 1) {
-         this.targetCallsign = "UNKNOWN";
-      } else {
-         targetCallsign = message.substring(1, msgSpc).trim().toUpperCase();
-      }
-      int msgNumberIdx = message.lastIndexOf('{');
-      this.messageNumber = "";
-      if (msgNumberIdx > -1) {
-         this.messageNumber = message.substring(msgNumberIdx + 1);
-         messageBody = message.substring(11, msgNumberIdx);
-      } else {
-         messageBody = message.substring(11);
-      }
-      String lcMsg = messageBody.toLowerCase();
-      if (lcMsg.startsWith("ack")) {
-         isAck = true;
-         this.messageNumber = messageBody.substring(3, messageBody.length());
-         this.messageBody = messageBody.substring(0, 3);
-      }
-      if (lcMsg.startsWith("rej")) {
-         isRej = true;
-         this.messageNumber = messageBody.substring(3, messageBody.length());
-         this.messageBody = messageBody.substring(0, 3);
-      }
-   }
+    /**
+     * @return the messageNumber
+     */
+    public String getMessageNumber() {
+        return messageNumber;
+    }
 
-   public MessagePacket(String targetCallsign, String messageBody, String messageNumber) {
-      this.messageBody = messageBody;
-      this.targetCallsign = targetCallsign;
-      this.messageNumber = messageNumber;
-      if (messageBody.equals("ack"))
-         isAck = true;
-      if (messageBody.equals("rej"))
-         isRej = true;
-      super.setDataTypeIdentifier(':');
-      super.type = APRSTypes.MESSAGE;
-   }
+    /**
+     * @param messageNumber the messageNumber to set
+     */
+    public void setMessageNumber(String messageNumber) {
+        this.messageNumber = messageNumber;
+    }
 
-   /**
-    * @return the messageBody
-    */
-   public String getMessageBody() {
-      return this.messageBody;
-   }
+    /**
+     * @return the targetCallsign
+     */
+    public String getTargetCallsign() {
+        return targetCallsign;
+    }
 
-   /**
-    * @param messageBody
-    *           the messageBody to set
-    */
-   public void setMessageBody(String messageBody) {
-      this.messageBody = messageBody;
-   }
+    /**
+     * @param targetCallsign the targetCallsign to set
+     */
+    public void setTargetCallsign(String targetCallsign) {
+        this.targetCallsign = targetCallsign;
+    }
 
-   /**
-    * @return the messageNumber
-    */
-   public String getMessageNumber() {
-      return messageNumber;
-   }
+	/**
+	 * @return the isAck
+	 */
+	public boolean isAck() {
+		return isAck;
+	}
 
-   /**
-    * @param messageNumber
-    *           the messageNumber to set
-    */
-   public void setMessageNumber(String messageNumber) {
-      this.messageNumber = messageNumber;
-   }
+	/**
+	 * @param isAck the isAck to set
+	 */
+	public void setAck(boolean isAck) {
+		this.isAck = isAck;
+	}
 
-   /**
-    * @return the targetCallsign
-    */
-   public String getTargetCallsign() {
-      return targetCallsign;
-   }
+	/**
+	 * @return the isRej
+	 */
+	public boolean isRej() {
+		return isRej;
+	}
 
-   /**
-    * @param targetCallsign
-    *           the targetCallsign to set
-    */
-   public void setTargetCallsign(String targetCallsign) {
-      this.targetCallsign = targetCallsign;
-   }
+	/**
+	 * @param isRej the isRej to set
+	 */
+	public void setRej(boolean isRej) {
+		this.isRej = isRej;
+	}
 
-   /**
-    * @return the isAck
-    */
-   public boolean isAck() {
-      return isAck;
-   }
+	
+    /** 
+     * @return String
+     */
+    @Override
+	public String toString() {
+		if (rawBytes != null)
+			return new String(rawBytes);
+		if ( this.messageBody.equals("ack") || this.messageBody.equals("rej")) {
+			return String.format(":%-9s:%s%s", this.targetCallsign, this.messageBody, this.messageNumber);
+		} else if (messageNumber.length() > 0) {
+			return String.format(":%-9s:%s{%s", this.targetCallsign, this.messageBody, this.messageNumber);
+		} else {
+			return String.format(":%-9s:%s", this.targetCallsign, this.messageBody);
+		}
+	}
 
-   /**
-    * @param isAck
-    *           the isAck to set
-    */
-   public void setAck(boolean isAck) {
-      this.isAck = isAck;
-   }
-
-   /**
-    * @return the isRej
-    */
-   public boolean isRej() {
-      return isRej;
-   }
-
-   /**
-    * @param isRej
-    *           the isRej to set
-    */
-   public void setRej(boolean isRej) {
-      this.isRej = isRej;
-   }
-
-   @Override
-   public String toString() {
-      if (rawBytes != null)
-         return new String(rawBytes);
-      if (this.messageBody.equals("ack") || this.messageBody.equals("rej")) {
-         return String.format(":%-9s:%s%s", this.targetCallsign, this.messageBody, this.messageNumber);
-      } else if (messageNumber.length() > 0) {
-         return String.format(":%-9s:%s{%s", this.targetCallsign, this.messageBody, this.messageNumber);
-      } else {
-         return String.format(":%-9s:%s", this.targetCallsign, this.messageBody);
-      }
-   }
+    @Override
+    public boolean hasFault() {
+        return hasFault;
+    }
 }
