@@ -5,286 +5,51 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.assertArrayEquals;
-import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.*;
  
 
-@DisplayName("Callsign Tests")
-class CallsignTest {
-	/* TODO Test SSID 0 */
+@DisplayName("Digipeater Tests")
+class DigipeaterTest {
 	/* TODO Test serialization */
+	/* TODO Fix code duplication */
 
 	@Nested
-	@DisplayName("Given regular APRS callsign with SSID")
-	class GivenRegularCallsignSsid {
-		final String callsign = "AB1CDE-12";
+	@DisplayName("Given unused, mixed-case APRS digipeater with SSID")
+	class GivenUnuseddDigipeaterSsid {
+		final String digipeater = "W1aw-15";
 
 		@Nested
 		@DisplayName("When instantiated as a class")
 		class WhenInstantiated {
-			Callsign object;
+			Digipeater object;
 
 			@BeforeEach
 			void setUp() {
-				object = new Callsign(callsign);
+				object = new Digipeater(digipeater);
 			}
 
 			@Test
 			@DisplayName("Then it should return the proper callsign")
 			void thenReturnCallsign() {
-				assertEquals("AB1CDE", object.getCallsign());
-			}
-
-			@Test
-			@DisplayName("Then it should return the proper SSID")
-			void thenReturnSsid() {
-				assertEquals("12", object.getSsid());
-			}
-
-			@Test
-			@DisplayName("Then it should return the proper string")
-			void thenReturnProperString() {
-				assertEquals("AB1CDE-12", object.toString());
-			}
-
-			@Test
-			@DisplayName("Then it should return correct AX25 data")
-			void thenReturnAX25() {
-				final byte[] expected = {
-					(byte)('A' << 1),
-					(byte)('B' << 1),
-					(byte)('1' << 1),
-					(byte)('C' << 1),
-					(byte)('D' << 1),
-					(byte)('E' << 1),
-					(byte)(0x60 | (12 << 1)),
-				};
-				assertEquals(7, object.toAX25().length);
-				assertArrayEquals(expected, object.toAX25());
-			}
-		}
-
-		@Nested
-		@DisplayName("When converted to bytes and back")
-		class WhenConverted {
-			Callsign object;
-
-			@BeforeEach
-			void setUp() {
-				object = new Callsign(
-					new Callsign(callsign).toAX25(), 0);
-			}
-
-			@Test
-			@DisplayName("Then it should return the proper callsign")
-			void thenReturnCallsign() {
-				assertEquals("AB1CDE", object.getCallsign());
-			}
-
-			@Test
-			@DisplayName("Then it should return the proper SSID")
-			void thenReturnSsid() {
-				assertEquals("12", object.getSsid());
-			}
-
-			@Test
-			@DisplayName("Then it should return the proper string")
-			void thenReturnProperString() {
-				assertEquals("AB1CDE-12", object.toString());
-			}
-
-			@Test
-			@DisplayName("Then it should return correct AX25 data")
-			void thenReturnAX25() {
-				final byte[] expected = {
-					(byte)('A' << 1),
-					(byte)('B' << 1),
-					(byte)('1' << 1),
-					(byte)('C' << 1),
-					(byte)('D' << 1),
-					(byte)('E' << 1),
-					(byte)(0x60 | (12 << 1)),
-				};
-				assertEquals(7, object.toAX25().length);
-				assertArrayEquals(expected, object.toAX25());
-			}
-		}
-	}
-
-	@Nested
-	@DisplayName("Given lowercase APRS callsign")
-	class GivenLowercaseCallsign {
-		final String callsign = "ab1cde";
-
-		@Nested
-		@DisplayName("When instantiated as a class")
-		class WhenInstantiated {
-			Callsign object;
-
-			@BeforeEach
-			void setUp() {
-				object = new Callsign(callsign);
-			}
-
-			@Test
-			@DisplayName("Then it should return the proper callsign")
-			void thenReturnCallsign() {
-				assertEquals("AB1CDE", object.getCallsign());
-			}
-
-			@Test
-			@DisplayName("Then it should return the proper SSID")
-			void thenReturnSsid() {
-				assertEquals("", object.getSsid());
-			}
-
-			@Test
-			@DisplayName("Then it should return the proper string")
-			void thenReturnProperString() {
-				assertEquals("AB1CDE", object.toString());
-			}
-
-			@Test
-			@DisplayName("Then it should return correct AX25 data")
-			void thenReturnAX25() {
-				final byte[] expected = {
-					(byte)('A' << 1),
-					(byte)('B' << 1),
-					(byte)('1' << 1),
-					(byte)('C' << 1),
-					(byte)('D' << 1),
-					(byte)('E' << 1),
-					(byte)(0x60 | (0 << 1)),
-				};
-				assertEquals(7, object.toAX25().length);
-				assertArrayEquals(expected, object.toAX25());
-			}
-		}
-
-		@Nested
-		@DisplayName("When converted to bytes and back")
-		class WhenConverted {
-			Callsign object;
-
-			@BeforeEach
-			void setUp() {
-				object = new Callsign(
-					new Callsign(callsign).toAX25(), 0);
-			}
-
-			@Test
-			@DisplayName("Then it should return the proper callsign")
-			void thenReturnCallsign() {
-				assertEquals("AB1CDE", object.getCallsign());
-			}
-
-			@Test
-			@DisplayName("Then it should return the proper SSID")
-			void thenReturnSsid() {
-				assertEquals("", object.getSsid());
-			}
-
-			@Test
-			@DisplayName("Then it should return the proper string")
-			void thenReturnProperString() {
-				assertEquals("AB1CDE", object.toString());
-			}
-
-			@Test
-			@DisplayName("Then it should return correct AX25 data")
-			void thenReturnAX25() {
-				final byte[] expected = {
-					(byte)('A' << 1),
-					(byte)('B' << 1),
-					(byte)('1' << 1),
-					(byte)('C' << 1),
-					(byte)('D' << 1),
-					(byte)('E' << 1),
-					(byte)(0x60 | (0 << 1)),
-				};
-				assertEquals(7, object.toAX25().length);
-				assertArrayEquals(expected, object.toAX25());
-			}
-		}
-	}
-
-	@Nested
-	@DisplayName("Given an APRS callsign")
-	class GivenCallsign {
-		Callsign object;
-
-		@BeforeEach
-		void setUp() {
-			object = new Callsign("W1AW-5");
-		}
-
-		@Nested
-		@DisplayName("When changing callsign")
-		class WhenCallsignChanged {
-			@BeforeEach
-			void setUp() {
-				object.setCallsign("NJ7P");
-			}
-
-			@Test
-			@DisplayName("Then it should return the new callsign")
-			void thenReturnNewCallsign() {
-				assertEquals("NJ7P", object.getCallsign());
-			}
-
-			@Test
-			@DisplayName("Then it should return the original SSID")
-			void thenReturnOriginalSsid() {
-				assertEquals("5", object.getSsid());
-			}
-
-			@Test
-			@DisplayName("Then it should return the proper string")
-			void thenReturnProperString() {
-				assertEquals("NJ7P-5", object.toString());
-			}
-
-			@Test
-			@DisplayName("Then it should return correct AX25 data")
-			void thenReturnAX25() {
-				final byte[] expected = {
-					(byte)('N' << 1),
-					(byte)('J' << 1),
-					(byte)('7' << 1),
-					(byte)('P' << 1),
-					(byte)(' ' << 1),
-					(byte)(' ' << 1),
-					(byte)(0x60 | (5 << 1)),
-				};
-				assertEquals(7, object.toAX25().length);
-				assertArrayEquals(expected, object.toAX25());
-			}
-		}
-
-		@Nested
-		@DisplayName("When changing ssid")
-		class WhenSsidChanged {
-			@BeforeEach
-			void setUp() {
-				object.setSsid("9");
-			}
-
-			@Test
-			@DisplayName("Then it should return the original callsign")
-			void thenReturnOriginalCallsign() {
 				assertEquals("W1AW", object.getCallsign());
 			}
 
 			@Test
-			@DisplayName("Then it should return the new SSID")
-			void thenReturnNewSsid() {
-				assertEquals("9", object.getSsid());
+			@DisplayName("Then it should return the proper SSID")
+			void thenReturnSsid() {
+				assertEquals("15", object.getSsid());
+			}
+
+			@Test
+			@DisplayName("Then it should return as unused")
+			void thenReturnUsed() {
+				assertFalse(object.isUsed(), "Should not be used");
 			}
 
 			@Test
 			@DisplayName("Then it should return the proper string")
 			void thenReturnProperString() {
-				assertEquals("W1AW-9", object.toString());
+				assertEquals("W1AW-15", object.toString());
 			}
 
 			@Test
@@ -297,7 +62,7 @@ class CallsignTest {
 					(byte)('W' << 1),
 					(byte)(' ' << 1),
 					(byte)(' ' << 1),
-					(byte)(0x60 | (9 << 1)),
+					(byte)(0*0x80 | 0x60 | (15 << 1)),
 				};
 				assertEquals(7, object.toAX25().length);
 				assertArrayEquals(expected, object.toAX25());
@@ -305,42 +70,219 @@ class CallsignTest {
 		}
 
 		@Nested
-		@DisplayName("When changing to mixed-case callsign")
-		class WhenCallsignChangedMixedCase {
+		@DisplayName("When converted to bytes and back")
+		class WhenConverted {
+			Digipeater object;
+
 			@BeforeEach
 			void setUp() {
-				object.setCallsign("N7lEm");
+				object = new Digipeater(
+					new Digipeater(digipeater).toAX25(), 0);
 			}
 
 			@Test
-			@DisplayName("Then it should return the new uppercase callsign")
-			void thenReturnNewCallsign() {
-				assertEquals("N7LEM", object.getCallsign());
+			@DisplayName("Then it should return the proper callsign")
+			void thenReturnCallsign() {
+				assertEquals("W1AW", object.getCallsign());
 			}
 
 			@Test
-			@DisplayName("Then it should return the original SSID")
-			void thenReturnOriginalSsid() {
-				assertEquals("5", object.getSsid());
+			@DisplayName("Then it should return the proper SSID")
+			void thenReturnSsid() {
+				assertEquals("15", object.getSsid());
+			}
+
+			@Test
+			@DisplayName("Then it should return as unused")
+			void thenReturnUsed() {
+				assertFalse(object.isUsed(), "Should not be used");
 			}
 
 			@Test
 			@DisplayName("Then it should return the proper string")
 			void thenReturnProperString() {
-				assertEquals("N7LEM-5", object.toString());
+				assertEquals("W1AW-15", object.toString());
 			}
 
 			@Test
 			@DisplayName("Then it should return correct AX25 data")
 			void thenReturnAX25() {
 				final byte[] expected = {
-					(byte)('N' << 1),
-					(byte)('7' << 1),
-					(byte)('L' << 1),
-					(byte)('E' << 1),
-					(byte)('M' << 1),
+					(byte)('W' << 1),
+					(byte)('1' << 1),
+					(byte)('A' << 1),
+					(byte)('W' << 1),
 					(byte)(' ' << 1),
-					(byte)(0x60 | (5 << 1)),
+					(byte)(' ' << 1),
+					(byte)(0*0x80 | 0x60 | (15 << 1)),
+				};
+				assertEquals(7, object.toAX25().length);
+				assertArrayEquals(expected, object.toAX25());
+			}
+		}
+	}
+
+	@Nested
+	@DisplayName("Given used, lower-case APRS digipeater")
+	class GivenUsedLowercaseDigipeater {
+		final String digipeater = "ab1cde*";
+
+		@Nested
+		@DisplayName("When instantiated as a class")
+		class WhenInstantiated {
+			Digipeater object;
+
+			@BeforeEach
+			void setUp() {
+				object = new Digipeater(digipeater);
+			}
+
+			@Test
+			@DisplayName("Then it should return the proper callsign")
+			void thenReturnCallsign() {
+				assertEquals("AB1CDE", object.getCallsign());
+			}
+
+			@Test
+			@DisplayName("Then it should return the proper SSID")
+			void thenReturnSsid() {
+				assertEquals("", object.getSsid());
+			}
+
+			@Test
+			@DisplayName("Then it should return as used")
+			void thenReturnUsed() {
+				assertTrue(object.isUsed(), "Should be used");
+			}
+
+			@Test
+			@DisplayName("Then it should return the proper string")
+			void thenReturnProperString() {
+				assertEquals("AB1CDE*", object.toString());
+			}
+
+			@Test
+			@DisplayName("Then it should return correct AX25 data")
+			void thenReturnAX25() {
+				final byte[] expected = {
+					(byte)('A' << 1),
+					(byte)('B' << 1),
+					(byte)('1' << 1),
+					(byte)('C' << 1),
+					(byte)('D' << 1),
+					(byte)('E' << 1),
+					(byte)(1*0x80 | 0x60 | (0 << 1)),
+				};
+				assertEquals(7, object.toAX25().length);
+				assertArrayEquals(expected, object.toAX25());
+			}
+		}
+
+		@Nested
+		@DisplayName("When converted to bytes and back")
+		class WhenConverted {
+			Digipeater object;
+
+			@BeforeEach
+			void setUp() {
+				object = new Digipeater(
+					new Digipeater(digipeater).toAX25(), 0);
+			}
+
+			@Test
+			@DisplayName("Then it should return the proper callsign")
+			void thenReturnCallsign() {
+				assertEquals("AB1CDE", object.getCallsign());
+			}
+
+			@Test
+			@DisplayName("Then it should return the proper SSID")
+			void thenReturnSsid() {
+				assertEquals("", object.getSsid());
+			}
+
+			@Test
+			@DisplayName("Then it should return as used")
+			void thenReturnUsed() {
+				assertTrue(object.isUsed(), "Should be used");
+			}
+
+			@Test
+			@DisplayName("Then it should return the proper string")
+			void thenReturnProperString() {
+				assertEquals("AB1CDE*", object.toString());
+			}
+
+			@Test
+			@DisplayName("Then it should return correct AX25 data")
+			void thenReturnAX25() {
+				final byte[] expected = {
+					(byte)('A' << 1),
+					(byte)('B' << 1),
+					(byte)('1' << 1),
+					(byte)('C' << 1),
+					(byte)('D' << 1),
+					(byte)('E' << 1),
+					(byte)(1*0x80 | 0x60 | (0 << 1)),
+				};
+				assertEquals(7, object.toAX25().length);
+				assertArrayEquals(expected, object.toAX25());
+			}
+		}
+	}
+
+	@Nested
+	@DisplayName("Given an APRS digipeater")
+	class GivenDigipeater {
+		Digipeater object;
+
+		@BeforeEach
+		void setUp() {
+			object = new Digipeater("W1AW-1");
+		}
+		@Nested
+		@DisplayName("When changing ssid")
+		class WhenUsedChanged {
+			@BeforeEach
+			void setUp() {
+				object.setUsed(true);
+			}
+
+			@Test
+			@DisplayName("Then it should return the proper callsign")
+			void thenReturnCallsign() {
+				assertEquals("W1AW", object.getCallsign());
+			}
+
+			@Test
+			@DisplayName("Then it should return the proper SSID")
+			void thenReturnNewSsid() {
+				assertEquals("1", object.getSsid());
+			}
+
+			@Test
+			@DisplayName("Then it should return as used")
+			void thenReturnUsed() {
+				assertTrue(object.isUsed(), "Should be used");
+			}
+
+			@Test
+			@DisplayName("Then it should return the proper string")
+			void thenReturnProperString() {
+				assertEquals("W1AW-1*", object.toString());
+			}
+
+			@Test
+			@DisplayName("Then it should return correct AX25 data")
+			void thenReturnAX25() {
+				final byte[] expected = {
+					(byte)('W' << 1),
+					(byte)('1' << 1),
+					(byte)('A' << 1),
+					(byte)('W' << 1),
+					(byte)(' ' << 1),
+					(byte)(' ' << 1),
+					(byte)(1*0x80 | 0x60 | (1 << 1)),
 				};
 				assertEquals(7, object.toAX25().length);
 				assertArrayEquals(expected, object.toAX25());
